+++ commands/moderation/warnlevel.js
@@ -1,0 +1,41 @@
+const Discord = require('discord.js');
+const fs = require("fs");
+const ms = require("ms");
+
+exports.run = async (client, message, args) => {
+    let warns = JSON.parse(fs.readFileSync("./warnings.json", "utf8"));
+    let user = message.mentions.users.first();
+    if (user === undefined) {
+        let userid = message.content.split(" ").slice(1, 2).join("")
+        user = await client.users.cache.get(userid)
+        if (user === undefined) {
+            message.channel.send("Please provide an actual mention or id!")
+        }
+    }
+
+    if(!warns[user.id]) warns[user.id] = {
+      warns: 0
+    };
+
+    const embed = new Discord.MessageEmbed()
+    .setColor(0xFFFF01)
+    .setTimestamp()
+    .addField('Action:', 'Warn Check')
+    .addField('User:', `${user.username}#${user.discriminator}`)
+    .addField('Number of warnings:', warns[`${user.id}, ${message.guild.id}`].warns)
+    .setTimestamp()
+    message.channel.send({embed});
+}
+
+exports.conf = {
+    enabled: true,
+    guildOnly: false,
+    aliases: [],
+    permLevel: 0
+  };
+  
+  exports.help = {
+    name: 'warnlevel',
+    description: 'Show how many warnings a user has',
+    usage: 'warnlevel [mention]'
+  };